docs: update module and function level docstrings
Update the docstring for the ConMan application by explaining what
Docker is and what ConMan is currently capable of doing with this
initial release.

Provide more context on how the entrypoint returns the exit code and
the time the error messages are logged.

Explain the correct usage of the logger module across a project with
relatively simple example.

Signed-off-by: Akshay Mestry (XAMES3) <xa@mes3.dev>